+++ lib/ModelMixins/ShadowMixin.ts
@@ -4,6 +4,7 @@ import Constructor from "../Core/Constructor";
 import Model from "../Models/Model";
 import { SelectableDimension } from "../Models/SelectableDimensions";
 import ShadowTraits, { Shadows } from "../Traits/TraitsClasses/ShadowTraits";
+import i18next from "i18next";
 
 function ShadowMixin<T extends Constructor<Model<ShadowTraits>>>(Base: T) {
   abstract class ShadowMixin extends Base {
@@ -31,12 +32,12 @@ function ShadowMixin<T extends Constructor<Model<ShadowTraits>>>(Base: T) {
     get shadowDimension(): SelectableDimension {
       return {
         id: "shadows",
-        name: "Shadows",
+        name: i18next.t("models.shadow.name"),
         options: [
-          { id: "NONE", name: "None" },
-          { id: "CAST", name: "Cast Only" },
-          { id: "RECEIVE", name: "Receive Only" },
-          { id: "BOTH", name: "Cast and Receive" }
+          { id: "NONE", name: i18next.t("models.shadow.options.none") },
+          { id: "CAST", name: i18next.t("models.shadow.options.cast") },
+          { id: "RECEIVE", name: i18next.t("models.shadow.options.receive") },
+          { id: "BOTH", name: i18next.t("models.shadow.options.both") }
         ],
         selectedId: this.shadows,
         disable: !this.showShadowUi,